backend/requirements.txt
Django==4.2.28
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
drf-spectacular==0.29.0
django-filter==25.1
django-extensions==4.1
psycopg2-binary==2.9.11
Markdown==3.9
gunicorn==20.1.0
